itt/requirements.txt
# =============================================================================
# Hybrid Prompt Optimization Pipeline — Dependencies
# =============================================================================
# Install: pip install -r requirements.txt

# --- Core ML / NLP ---
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0

# --- Image Generation ---
diffusers>=0.25.0
Pillow>=10.0.0

# --- Visualization & Analysis ---
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
numpy>=1.24.0

# --- Statistical Testing ---
scipy>=1.11.0

# --- LLM Providers (install the one you need) ---
openai>=1.0.0          # For OpenAI and Groq (OpenAI-compatible)
# anthropic>=0.18.0    # Uncomment if using Claude
